Move `toOutgoingContent` to `DavUtils`
